Overwrites original output piping test
